refactor: extract file transfer operations to dedicated service
- Created IFileTransferService interface with ExecuteUploadsAsync/ExecuteDownloadsAsync
- Implemented FileTransferService with parallel upload/download logic
- Moved CreateUploadTasks, CreateDownloadTasks, and batch handling methods from SyncEngine
- Reduced SyncEngine from ~1062 to ~850 lines (-212 lines)
- Reduced SyncEngine constructor from 11 to 12 parameters (one service replaces multiple low-level dependencies)
- Updated SyncEngine to inject and delegate to FileTransferService
- Added FileTransferService DI registration in AppHost
- Updated SyncEngineShould tests to mock IFileTransferService

This addresses architectural concern of SyncEngine doing too much by separating file transfer orchestration into a focused, testable service.
