+++ src/AStar.Dev.OneDrive.Client/ AppHost.cs
@@ -132,6 +132,7 @@ public static class AppHost
         _ = services.AddScoped<IRemoteChangeDetector, RemoteChangeDetector>();
         _ = services.AddScoped<IConflictResolver, ConflictResolver>();
         _ = services.AddScoped<ISyncEngine, SyncEngine>();
+        _ = services.AddScoped<IFileTransferService, FileTransferService>();
         _ = services.AddScoped<IDebugLogger, DebugLoggerService>();
         _ = services.AddScoped<IDeltaPageProcessor, DeltaPageProcessor>();
         _ = services.AddScoped<ISyncRepository, EfSyncRepository>();
